Move test and black
